Rename topic information header file.
