+++ ecal/core/include/ecal/types/ecal_topic_information.h
