+++ ecal/core/src/ecal_descgate.h
@@ -24,7 +24,7 @@
 #pragma once
 
 #include <ecal/ecal_util.h>
-#include <ecal/ecal_types.h>
+#include <ecal/types/ecal_topic_information.h>
 
 #include "ecal_global_accessors.h"
 #include "ecal_def.h"
