+++ ecal/core/src/readwrite/ecal_reader.h
@@ -25,7 +25,7 @@
 
 #include <ecal/ecal.h>
 #include <ecal/ecal_callback.h>
-#include <ecal/ecal_types.h>
+#include <ecal/types/ecal_topic_information.h>
 
 #ifdef _MSC_VER
 #pragma warning(push, 0) // disable proto warnings
