+++ ecal/core/src/readwrite/ecal_writer.h
@@ -26,7 +26,7 @@
 #include <ecal/ecal_callback.h>
 #include <ecal/ecal_payload_writer.h>
 #include <ecal/ecal_tlayer.h>
-#include <ecal/ecal_types.h>
+#include <ecal/types/ecal_topic_information.h>
 
 #include "ecal_def.h"
 #include "ecal_expmap.h"
